+++ src/modules/auth/auth.service.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { AuthService } from './auth.service';
-import { UserDocument } from '../user/user.document';
+import { UserDocument } from '../../mongodb/documents/user.document';
 import { JwtManagerService } from '../jwt-manager/jwt-manager.service';
 import { UserService } from '../user/user.service';
 import { LoggerService } from '../logger/logger.service';
@@ -8,8 +8,8 @@ import { JwtService } from '@nestjs/jwt';
 import { NotFoundException } from '../../exceptions/not-found.exception';
 import { REDIS_CLIENT } from '../redis/redis.constants';
 import { MailManagerService } from '../mail-manager/mail-manager.service';
-import { UserRepository } from '../../repositories/user.repository';
-import { UserActionRepository } from '../../repositories/user.action.repository';
+import { UserRepository } from '../../mongodb/repositories/user.repository';
+import { UserActionRepository } from '../../mongodb/repositories/user.action.repository';
 
 // https://betterprogramming.pub/testing-controllers-in-nestjs-and-mongo-with-jest-63e1b208503c
 // https://stackoverflow.com/questions/74110962/please-make-sure-that-the-argument-databaseconnection-at-index-0-is-available